Struct opentelemetry_prometheus::ExporterBuilder

pub struct ExporterBuilder { /* fields omitted */ }

Configuration for the prometheus exporter.

Implementations

impl ExporterBuilder

pub fn with_resource(self, resource: Resource) -> Self

Set the resource to be associated with all Meters for this exporter

pub fn with_cache_period(self, period: Duration) -> Self

Set the period which a recently-computed result will be returned without gathering metric data again.

If the period is zero, caching of the result is disabled (default).

pub fn with_default_summary_quantiles(self, quantiles: Vec<f64>) -> Self

Set the default summary quantiles to be used by exported prometheus histograms

pub fn with_default_histogram_boundaries(self, boundaries: Vec<f64>) -> Self

Set the default boundaries to be used by exported prometheus histograms

pub fn with_host(self, host: String) -> Self

Set the host for the prometheus exporter

pub fn with_port(self, port: u16) -> Self

Set the port for the prometheus exporter

pub fn with_registry(self, registry: Registry) -> Self

Set the prometheus registry to be used by this exporter

pub fn with_aggregator_selector<T>(self, aggregator_selector: T) -> Self where
T: AggregatorSelector + Send + Sync + 'static,

Set the aggregation selector for the prometheus exporter

pub fn try_init(self) -> Result<PrometheusExporter, MetricsError>

Sets up a complete export pipeline with the recommended setup, using the recommended selector and standard processor.

pub fn init(self) -> PrometheusExporter

Sets up a complete export pipeline with the recommended setup, using the recommended selector and standard processor.

Panics

This panics if the exporter cannot be registered in the prometheus registry.